eliminate some warnings, but not all of them
git-svn-id: file:///local/work/git-test/fastprotect.svn/trunk@9617 eaa77fc2-5f71-4410-aed7-0df95d92e94a

+++ src/osgEarthDrivers/engine_quadtree/TileModelCompiler.cpp
@@ -998,7 +998,7 @@ namespace
 
             if (orig_i < 0)
             {
-                if (skirtBreaks.back() != skirtVerts->size())
+                if (skirtBreaks.back() != (int)skirtVerts->size())
                     skirtBreaks.push_back(skirtVerts->size());
             }
             else
@@ -1041,7 +1041,7 @@ namespace
             int orig_i = d.indices[r*d.numCols+(d.numCols-1)];
             if (orig_i < 0)
             {
-                if (skirtBreaks.back() != skirtVerts->size())
+                if (skirtBreaks.back() != (int)skirtVerts->size())
                     skirtBreaks.push_back(skirtVerts->size());
             }
             else
@@ -1084,7 +1084,7 @@ namespace
             int orig_i = d.indices[(d.numRows-1)*d.numCols+c];
             if (orig_i < 0)
             {
-                if (skirtBreaks.back() != skirtVerts->size())
+                if (skirtBreaks.back() != (int)skirtVerts->size())
                     skirtBreaks.push_back(skirtVerts->size());
             }
             else
@@ -1127,7 +1127,7 @@ namespace
             int orig_i = d.indices[r*d.numCols];
             if (orig_i < 0)
             {
-                if (skirtBreaks.back() != skirtVerts->size())
+                if (skirtBreaks.back() != (int)skirtVerts->size())
                     skirtBreaks.push_back(skirtVerts->size());
             }
             else
